requirements.txt
Django==1.6.1
Pygments==1.6
djangorestframework==2.3.13
argparse==1.2.1
dj-database-url==0.3.0
psycopg2==2.5.3
wsgiref==0.1.2
